Restore previous (and more correct) constraint match logic
Update the unit tests to include tests we expect to fail and tests we
expect to succeed when tested for a match.

Some tests have captured possible spec non-conformance when matching a
version constraint including a prerelease tag and a version with a
prerelease tag that is not equal to but greater than and within the
bounds of the constraint. However, the behavior of the matching logic
maintains the important invariants and also maintains the invariant
that no version constraint without a prerelease tag matches a version
_with_ a prerelease tag.